SO-4640: remove unnecessary DatastoreOperationLockException
Use OperationLockException instead.
Simplify lock exception handling.
